updated the switch handler and some warnings
updated the switch handler with an delay, like the original code intended.

added a switch state handler in main.c that does stuff according to the button that has been pressed, and filtered in switch.c

switch 1 now iterates through the programStates.

switch 2 now iterated through the moods

both switches pressed initiates the test sequence.

Fixed some warnings in tempsensor.

removed uneccesary comments

+++ tempusludicus/Core/Inc/tempSensor.h
@@ -11,7 +11,7 @@
 #include <MKL25Z4.h>
 
 void init_adc_lm35(void);
-uint16_t read_adc_lm35(void);
+uint32_t read_adc_lm35(void);
 
 float calculate_temperature_from_lm35(uint16_t adc_result);
 void addTemperatureToBuffer(float temperature);
